- cmake: removed check for doxygen since it is not used - eliminated scanf/printf warnings related to %llu on 64bit architecture by adding casts to (long long unsigned) - IpfixPrinter prints IE id
git-svn-id: file:///Users/braun/svn/vermont/branches/vermont/new-template@2247 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ src/modules/analysis/AutoFocus.cpp
@@ -342,7 +342,7 @@ void AutoFocus::metalist()
 			sprintf(num,"%25s",(*iter)->global.c_str());
 			locl.append(num);
 			locl.append("\t");
-			sprintf(num,"%10llu",data);
+			sprintf(num,"%10llu",(long long unsigned)data);
 			locl.append(num);
 			locl.append(" \t");
 			percentage = (double) (data*100) / (double) (*iter)->numTotal;
